utils: Move Git object IO helpers
The Git utility layer provides command execution, index plumbing,
repository path lookup, and blob/tree object IO from one module.

The project lacks a named boundary for streaming Git object content. That
makes data and batch modules depend on the broad command utility when they
only need hash-object, cat-file, or ls-tree blob reads.

This commit addresses that boundary gap by moving GitTreeBlob plus blob
creation, blob reading, batch blob reading, and tree blob listing into
utils.git_object_io, then updating source callers to import that object IO
module directly.

Subsequent commits will update tests and add architecture guards for the
new Git object IO boundary.
