+++ src/git_stage_batch/commands/batch_source/binary_file_actions.py
@@ -13,10 +13,10 @@ from ...core.buffer import (
 from ...data.file_modes import apply_git_file_mode, detect_file_mode_in_commit
 from ...utils.repository_buffers import load_git_object_as_buffer
 from ...utils.git import (
-    create_git_blob,
     get_git_repository_root_path,
     git_update_index,
 )
+from ...utils.git_object_io import create_git_blob
 
 
 class BinaryWorktreeAction(Enum):
