+++ src/git_stage_batch/commands/batch_source/text_file_actions.py
@@ -12,7 +12,8 @@ from ...core.buffer import (
 from ...core.text_lifecycle import TextFileChangeType, normalized_text_change_type
 from ...data.file_modes import apply_git_file_mode
 from ...staging.operations import update_index_with_blob_buffer
-from ...utils.git import create_git_blob, get_git_repository_root_path, git_update_index
+from ...utils.git import get_git_repository_root_path, git_update_index
+from ...utils.git_object_io import create_git_blob
 
 
 def stage_text_file_to_index(
