+++ src/git_stage_batch/data/batch_sources.py
@@ -19,17 +19,16 @@ from ..exceptions import CommandError
 from ..i18n import _
 from ..utils.file_io import read_file_paths_file, read_text_file_contents, write_text_file_contents
 from ..utils.git import (
-    create_git_blob,
     get_git_repository_root_path,
     git_commit_tree,
     git_read_tree,
     git_update_index,
     git_write_tree,
-    list_git_tree_blobs,
     run_git_command,
     temp_git_index,
     update_git_refs,
 )
+from ..utils.git_object_io import create_git_blob, list_git_tree_blobs
 from ..utils.journal import log_journal
 from ..utils.paths import (
     get_abort_head_file_path,
