+++ src/git_stage_batch/data/undo.py
@@ -24,7 +24,6 @@ from . import undo_worktree as _undo_worktree
 from ..exceptions import CommandError
 from ..i18n import _
 from ..utils.git import (
-    create_git_blob,
     GitIndexEntryUpdate,
     get_git_repository_root_path,
     git_commit_tree,
@@ -35,6 +34,7 @@ from ..utils.git import (
     temp_git_index,
     update_git_refs,
 )
+from ..utils.git_object_io import create_git_blob
 from ..utils.paths import (
     get_batches_directory_path,
     get_session_directory_path,
