+++ src/git_stage_batch/data/undo_worktree.py
@@ -9,11 +9,11 @@ from typing import Any
 
 from ..core.buffer import LineBuffer
 from ..utils.git import (
-    create_git_blob,
     get_git_repository_root_path,
     GitIndexEntryUpdate,
     run_git_command,
 )
+from ..utils.git_object_io import create_git_blob
 
 
 def changed_worktree_paths() -> list[str]:
