+++ src/git_stage_batch/utils/repository_buffers.py
@@ -9,10 +9,9 @@ from collections.abc import Iterator
 from ..core.buffer import LineBuffer
 from ..utils.git import (
     get_git_repository_root_path,
-    list_git_tree_blobs,
-    read_git_blob,
     stream_git_command,
 )
+from ..utils.git_object_io import list_git_tree_blobs, read_git_blob
 
 
 def load_git_blob_as_buffer(blob_sha: str) -> LineBuffer:
